ethtool: fbnic: ethtool dump parser
This patch adds support for parsing the register dump for fbnic.

The patch is composed of several register sections, and each of these
sections is dumped lineraly except for the RPC_RAM section which is handled
differently.

For each of the sections, we dump register name, its value, the bit mask
of any subfields within that register, the name of the subfield, and the
corresponding value.

Furthermore, there may be unused blocks within a section; we skip such
blocks while dumping registers linearly.

Validation:
- Validate patch applies to master without any warning
- Validate 'ethtool -d' for net-next branch generates ascii dump
	$ uname -r
	  6.14.0-0_fbk701_rc0_429_g8e5edf971d0
	$ ./ethtool -d eth0 > /tmp/fbnic_ascii_dump

- Verify patch compiles without any warning on 32bit
  `make clean && ./configure CFLAGS="-m32" && make && make check`
   =======================================================================
   Testsuite summary for ethtool 6.14
   =======================================================================
   # TOTAL: 1
   # PASS:  1
   # SKIP:  0
   # XFAIL: 0
   # FAIL:  0
   # XPASS: 0
   # ERROR: 0
   =======================================================================

Signed-off-by: Mohsin Bashir <mohsin.bashr@gmail.com>