module_common: always print per-lane status in JSON
The JSON output type changes when loss of signal / fault is
detected. When there is no problem we print single bool, eg:

  "rx_loss_of_signal": false,

but when there's a problem we print an array:

  "rx_loss_of_signal": ["No", "Yes", "No", "No"],

This appears to be a mirror of the human-readable output,
but it's a pain to parse / unmarshall for user space.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>
(cherry picked from commit 7ca78b77af7482eb189d15e26c3464ddcb5b5d58)